+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+1.1.0
+=====
+
+*   Removed the `sort` option, the items are now automatically sorted: first desc by priority. 
+    If the priority is the same, then asc by label.
+
+
 1.0.1
 =====